refactor: Move streching to position service
