+++ src/app/services/util/position.transformation.service.spec.ts
@@ -105,6 +105,7 @@ describe("PositionTransformationService", () => {
 
     positionTransformationService = new PositionTransformationService(
       trainrunSectionService,
+      trainrunService,
       nodeService,
       noteService,
       uiInteractionService,
