+++ src/app/view/editor-main-view/data-views/d3.utils.spec.ts
@@ -132,12 +132,13 @@ describe("3d.Utils.tests", () => {
     );
 
     const positionTransformationService = new PositionTransformationService(
-      trainrunSectionService,
-      nodeService,
-      noteService,
-      uiInteractionService,
-      viewportCullService,
-    );
+          trainrunSectionService,
+          trainrunService,
+          nodeService,
+          noteService,
+          uiInteractionService,
+          viewportCullService,
+        );
 
     const controller = new EditorMainViewComponent(
       nodeService,
