+++ src/app/view/editor-main-view/data-views/notes.view.spec.ts
@@ -132,12 +132,13 @@ describe("Notes-View", () => {
     );
 
     const positionTransformationService = new PositionTransformationService(
-      trainrunSectionService,
-      nodeService,
-      noteService,
-      uiInteractionService,
-      viewportCullService,
-    );
+          trainrunSectionService,
+          trainrunService,
+          nodeService,
+          noteService,
+          uiInteractionService,
+          viewportCullService,
+        );
 
     const controller = new EditorMainViewComponent(
       nodeService,
