+++ src/app/view/editor-main-view/data-views/trainrunsections.view.spec.ts
@@ -134,12 +134,13 @@ describe("TrainrunSection-View", () => {
     );
 
     const positionTransformationService = new PositionTransformationService(
-      trainrunSectionService,
-      nodeService,
-      noteService,
-      uiInteractionService,
-      viewportCullService,
-    );
+          trainrunSectionService,
+          trainrunService,
+          nodeService,
+          noteService,
+          uiInteractionService,
+          viewportCullService,
+        );
 
     const controller = new EditorMainViewComponent(
       nodeService,
